refactor(pd): testnet logic moved to module
Shoves things around a bit and fleshes out the docstrings.
The `testnet generate` logic has been largely superseded by the
config-munging in the k8s deploy scripts, but timeboxed the refactor to
focus on organization.
